volatile: Add MMIO access via Hal
Add additional branches in the volread!() and volwrite!() macros
that take the Hal as an additional type parameter and use its
mmio_read() and mmio_write() functions.

Once all code is converted to using the Hal for mmio, the currently
existing code will be obsolete and can be removed. Until then, let's
keep it around.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>